+++ app/src/main/java/com/voltek/newsfeed/domain/usecase/BaseUseCase.kt
@@ -1,4 +1,4 @@
-package com.voltek.newsfeed.domain.use_case
+package com.voltek.newsfeed.domain.usecase
 
 import com.voltek.newsfeed.utils.SubscriptionsHolder
 import io.reactivex.Observable
