+++ app/src/main/java/com/voltek/newsfeed/domain/usecase/articles/GetArticlesUseCase.kt
@@ -1,10 +1,10 @@
-package com.voltek.newsfeed.domain.use_case.articles
+package com.voltek.newsfeed.domain.usecase.articles
 
 import com.voltek.newsfeed.domain.repository.ArticlesRepository
 import com.voltek.newsfeed.domain.repository.NewsSourcesRepository
-import com.voltek.newsfeed.domain.use_case.BaseUseCase
-import com.voltek.newsfeed.domain.use_case.Parameter
-import com.voltek.newsfeed.domain.use_case.Result
+import com.voltek.newsfeed.domain.usecase.BaseUseCase
+import com.voltek.newsfeed.domain.usecase.Parameter
+import com.voltek.newsfeed.domain.usecase.Result
 import com.voltek.newsfeed.presentation.entity.ArticleUI
 import io.reactivex.Observable
 import io.reactivex.Scheduler
