+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/list/ListPresenter.kt
@@ -2,9 +2,9 @@ package com.voltek.newsfeed.presentation.ui.list
 
 import com.arellomobile.mvp.InjectViewState
 import com.voltek.newsfeed.Logger
-import com.voltek.newsfeed.domain.use_case.Parameter
-import com.voltek.newsfeed.domain.use_case.articles.GetArticlesUseCase
-import com.voltek.newsfeed.domain.use_case.news_sources.NewsSourcesUpdatesUseCase
+import com.voltek.newsfeed.domain.usecase.Parameter
+import com.voltek.newsfeed.domain.usecase.articles.GetArticlesUseCase
+import com.voltek.newsfeed.domain.usecase.newssources.NewsSourcesUpdatesUseCase
 import com.voltek.newsfeed.presentation.base.BasePresenter
 import com.voltek.newsfeed.presentation.base.Event
 import com.voltek.newsfeed.presentation.navigation.command.CommandOpenArticleDetailsScreen
@@ -20,6 +20,7 @@ class ListPresenter(
         private val newsSourcesChanges: NewsSourcesUpdatesUseCase
 ) : BasePresenter<ListView>() {
 
+    // Holds current model through full presenter lifecycle
     private val model: ListModel = ListModel { viewState.render(it as ListModel) }
 
     init {
@@ -31,6 +32,7 @@ class ListPresenter(
         loadArticles()
     }
 
+    // View event presenter about events using this method
     override fun event(event: Event) {
         when (event) {
             is Event.OpenArticleDetails -> router.execute(CommandOpenArticleDetailsScreen(event.article))
@@ -84,7 +86,7 @@ class ListPresenter(
                     loadArticles()
                 },
                 Consumer {
-                    Logger.e(it)
+                    it.printStackTrace()
                 },
                 Action {}
         )
