+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/newssources/NewsSourcesPresenter.kt
@@ -1,9 +1,9 @@
 package com.voltek.newsfeed.presentation.ui.newssources
 
 import com.arellomobile.mvp.InjectViewState
-import com.voltek.newsfeed.domain.use_case.Parameter
-import com.voltek.newsfeed.domain.use_case.news_sources.EnableNewsSourceUseCase
-import com.voltek.newsfeed.domain.use_case.news_sources.NewsSourcesUseCase
+import com.voltek.newsfeed.domain.usecase.Parameter
+import com.voltek.newsfeed.domain.usecase.newssources.EnableNewsSourceUseCase
+import com.voltek.newsfeed.domain.usecase.newssources.NewsSourcesUseCase
 import com.voltek.newsfeed.presentation.base.BasePresenter
 import com.voltek.newsfeed.presentation.base.Event
 import com.voltek.newsfeed.presentation.entity.SourceUI
