+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/splash/SplashPresenter.kt
@@ -3,8 +3,8 @@ package com.voltek.newsfeed.presentation.ui.splash
 import com.arellomobile.mvp.InjectViewState
 import com.voltek.newsfeed.Logger
 import com.voltek.newsfeed.domain.exception.NoNewsSourcesSelectedException
-import com.voltek.newsfeed.domain.use_case.Parameter
-import com.voltek.newsfeed.domain.use_case.news_sources.NewsSourcesUseCase
+import com.voltek.newsfeed.domain.usecase.Parameter
+import com.voltek.newsfeed.domain.usecase.newssources.NewsSourcesUseCase
 import com.voltek.newsfeed.presentation.base.BasePresenter
 import com.voltek.newsfeed.presentation.base.Event
 import com.voltek.newsfeed.presentation.entity.SourceUI
