+++ app/src/test/java/com/voltek/newsfeed/domain/usecase/newssources/NewsSourcesUpdatesUseCaseTest.kt
@@ -1,11 +1,11 @@
-package com.voltek.newsfeed.domain.use_case.news_sources
+package com.voltek.newsfeed.domain.usecase.newssources
 
 import com.nhaarman.mockito_kotlin.times
 import com.nhaarman.mockito_kotlin.verify
 import com.nhaarman.mockito_kotlin.whenever
 import com.voltek.newsfeed.domain.repository.NewsSourcesRepository
-import com.voltek.newsfeed.domain.use_case.BaseUseCaseTest
-import com.voltek.newsfeed.domain.use_case.Parameter
+import com.voltek.newsfeed.domain.usecase.BaseUseCaseTest
+import com.voltek.newsfeed.domain.usecase.Parameter
 import io.reactivex.functions.Action
 import io.reactivex.functions.Consumer
 import io.reactivex.subjects.PublishSubject
