+++ app/src/test/java/com/voltek/newsfeed/presentation/ui/newssources/NewsSourcesPresenterTest.kt
@@ -4,8 +4,8 @@ import com.nhaarman.mockito_kotlin.any
 import com.nhaarman.mockito_kotlin.times
 import com.nhaarman.mockito_kotlin.verify
 import com.voltek.newsfeed.MockData
-import com.voltek.newsfeed.domain.use_case.news_sources.EnableNewsSourceUseCase
-import com.voltek.newsfeed.domain.use_case.news_sources.NewsSourcesUseCase
+import com.voltek.newsfeed.domain.usecase.newssources.EnableNewsSourceUseCase
+import com.voltek.newsfeed.domain.usecase.newssources.NewsSourcesUseCase
 import com.voltek.newsfeed.presentation.base.Event
 import com.voltek.newsfeed.presentation.ui.BasePresenterTest
 import org.junit.Before
